+++ NEWS.md
@@ -1,1 +1,7 @@
-# r.pkg.template 0.0.0.91
+# r.pkg.template 0.1.0
+### New features
+* Add an initializer script.
+### Enhancements
+* Documentation on how to use the initialize a package.
+### Bug fixes
+* None.